+++ src/main/java/com/example/coffeeshop/Services/TuotteenHallintaService.java
@@ -71,5 +71,7 @@ public class TuotteenHallintaService {
         tuoteRepository.save(tuote);
 
     }
+
+    
     
 }
