+++ src/main/resources/static/css/kahvilaite.css
@@ -1,0 +1,26 @@
+.background-product {
+    background-color: rgba(0, 0, 0, 0.68);
+    
+    
+}
+
+.px-3 {
+    padding: 5em;
+}
+
+.detail.image {
+    max-width: 100%;
+    
+}
+
+img {
+   width: 75%;
+  
+}
+.detail-img {
+    
+}
+p {
+
+}
+
