+++ src/main/resources/static/css/kahvilaitteet.css
@@ -1,0 +1,8 @@
+
+
+.card.shadow-sm{
+   /*  border: 2px solid black; */
+}
+img {
+    padding:1em;
+}
