+++ src/main/resources/templates/fragments/head.html
@@ -12,6 +12,8 @@
     <link rel="stylesheet" type="text/css" th:href="@{/css/tuotteen-muokkaus.css}">
     <link rel="stylesheet" type="text/css" th:href="@{/css/registration.css}">
     <link rel="stylesheet" type="text/css" th:href="@{/css/footer.css}">
+    <link rel="stylesheet" type="text/css" th:href="@{/css/kahvilaite.css}">
+    <link rel="stylesheet" type="text/css" th:href="@{/css/kahvilaitteet.css}">
     
     <title th:text="${title}"></title>
     <script src="https://kit.fontawesome.com/0f2be22396.js" crossorigin="anonymous"></script>
