+++ src/main/resources/templates/test.html
@@ -7,7 +7,6 @@
     <link rel="stylesheet" th:href="@{/css/cover.css}">
     <title>Test</title>
     <link href="/docs/5.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-
 </head>
 <body class="d-flex h-100 text-center text-white bg-dark" cz-shortcut-listen="true">
     
@@ -30,15 +29,6 @@
           <a href="#" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Learn more</a>
         </p>
       </main>
-    
-      <footer class="mt-auto text-white-50">
-        <p>Cover template for <a href="https://getbootstrap.com/" class="text-white">Bootstrap</a>, by <a href="https://twitter.com/mdo" class="text-white">@mdo</a>.</p>
-      </footer>
-    </div>
-    
-    
-        
-      
     
     </body>
 </html>
